refactor: api 요청 훅 디렉토리 변경
